[jinmei-inmem-ng] encode node names as label sequences, not Name objects.
insert/find are now label-sequence based.  this means compare() can return
NONE, and it should be handled accordingly.  same for previousNode().
Node::name_ is now gone.

in the unit test tree, there is now a single "." node.  this changes
some assumptions of the tests, and they should be adjusted accordingly.