[1899] Break up UNION SQL query into two and merge manually
We also add a new index nsec3_byowner_and_rdtype, but don't bump
the SQL schema (the plan is to merge #1756 which bumps schema
version and #1899 together so that two schema version changes are
not necessary).

Note that this still is buggy code, and it has always been buggy
because it tries to merge rname (reversed name) from the records
table with owner column in the nsec3 table. So you have results
out of order. There was no unit test to check if both sides of
the UNION were being merged properly before.

We'd have to fix this by adding nsec3.rowner column.