[stephen-innodb] Use InnoDB engine for asynchronous MySQL DHCP Tests
All MySQL tests use the InnoDB engine: autocommit is enabled for a
synchronous test and disabled for an asynchronous one.  In both
cases, a final commit is made just prior to disconnecting from the
database.

Added timers to measure the time taken to connect and disconnect (the
latter being an upper limit on the time taken to do a commit).

Converted some variables to "unsigned long" to avoid warnings from
g++ compiler.

Fixed bug whereby 100 leases were created but only 99 deleted.