[2776] Let the TCP server accept batches of messages
Each batch is preceded by 4-byte length of the batch. Then the messages
are put one after another, with the usual 2-byte length of message.

This should allow for less reads and less context switches in the
following experiment.

As this is experiment, no tests are updated or written. The old ones,
obviously, don't work, as this is backwards-incompatible change. Hoping
it just works, if not, it'll be debugged later when the other part is
written.

Also, the code style is probably not clean either, comments and docs are
missing, etc.

This will work for auth, but probably not for the resolver (or, it
would, but not well, as it would block for one message before starting
another). This is not an issue with auth server, as all the lookups are
synchronous.