[2874] Skip coroutines benchmark if no lib
Skip the coroutines resolver benchmark during the compilation, if the
boost coroutines library is not available.

The -lboost_context is needed to link it. It is not clear if any more
flags (like paths) need to be used on some systems, this is just
guesswork to make it work locally.

Some better detection or configure switch TBD here.