[2989] More locking
Add more locking, the original code didn't look safe (there are some
places that are still doubtful, but the goal of this ticket is not to go
through them; this is just around the client lists added here).

The locks might be wider & more frequent than actually needed, but it's
better to err on the safer side. The code is not performance sensitive,
so it should not matter.