[360] Check for socket directory being too long
At configure time check the length of the path where the message queue
socket endpoint will be created and produce an error message if it is
too long. Add the --with-socket-directory configure option to allow the
user to specify a shorter path if the default based on $prefix ends up
being too long.