[2904] Change test port from 4780XX to 5617XX and use IPv4
A change was made in the past to get around an Apple service that runs
onf port 47807. Part of that change was to change from using 127.0.0.1
and start using ::1, thus changing things to use IPv6, which the Apple
service did not bind to. Reverse that change and use ports higher up
in the RFC6335 "dynamic ports" range. If we want to test IPv6 then we
should have specific tests for that. Not all systems will have IPv6
support enabled and the current tests will all fail on such systems.