[3243] Added bootstrap.sh to cater for serial tests
In automake 1.13 and later the default test harness was changed to be
the new parallel test harness. This is undesirable in certain situations
where we may want the test output sent to stdout versus to an individual
log file (for example in the build farm). bootstrap.sh is now required
in order to generate a special file that is included by configure.ac to
controll which test harness should be used, if the version of automake
being used supports this. If you want to build using the serial test
harness invoke bootstrap.sh with the single argument "serial".