[2367_v2] Shorten help message to most common options
Rather than display the list of all components, which may lead users
to expect they can configure and build just a single component, adjust
the help string for --enable-components to just list the meta-names for
DNS, DHCP and common code. All of the other component names still
exist, so fine-grained control is still possible, but that is not the
intent behind this option.